+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.26.0] - 2025-11-21
+
+### Added
+
+- Import Toast Component
+
+### Fixed
+
+- Toast styles
+
 ## [1.25.2] - 2025-11-12
 
 ### Fixed